refactor: migrate bing library to bing_search
