+++ cogniq/bots/bing_search/__init__.py
@@ -1,1 +1,1 @@
-from .bing_search import start
+from .start import start
